@@ -1,28 +1,45 @@
 [project]
 name = "Turbacz"
-version = "0.2.0"
+version = "1.0.0"
 description = "IoT logic and user interface"
-authors = [{ name = "DrDubel", email = "antek.suszko@gmail.com" }]
-requires-python = ">=3.10,<4"
-license = "Apache 2.0"
+license = "Apache-2.0"
+authors = [
+    { name = "DrDubel", email = "antek.suszko@gmail.com" }
+]
+requires-python = ">=3.11"
+
 dependencies = [
-    "fastapi",
-    "fastapi-mqtt",
-    "uvicorn",
-    "Authlib",
-    "itsdangerous",
-    "httpx",
-    "aioconsole",
-    "uvloop",
-    "aioprometheus",
-    "wsproto",
+    "fastapi>=0.121.2",
+    "fastapi-mqtt>=2.2.0",
+    "uvicorn<0.38.0",
+    "Authlib>=1.6.5",
+    "itsdangerous>=2.2.0",
+    "httpx>=0.28.1",
+    "aioconsole>=0.8.2",
+    "uvloop>=0.22.1",
+    "aioprometheus>=23.12.0",
+    "wsproto>=1.3.1",
+    "unique-namer>=1.6.2",
+    "python-multipart>=0.0.21",
+    "python-jose[cryptography]>=3.5.0",
+    "sentry-sdk[fastapi]>=2.48.0",
 ]
 
 [project.scripts]
-start = "turbacz.main:start"
+start = "czupel.main:start"
 
-[dependency-groups]
+[project.optional-dependencies]
 dev = [
-    "black",
-    "isort",
+    "black>=25.11.0",
+    "isort>=7.0.0",
 ]
+
+[tool.setuptools]
+packages = ["turbacz"]
+
+[tool.setuptools.package-data]
+turbacz = ["../static/**"]
+
+[build-system]
+requires = ["setuptools", "wheel"]
+build-backend = "setuptools.build_meta"
