@@ -1,29 +1,68 @@
+import json
 import logging
 import os
-from pickle import dump, load
-from secrets import token_urlsafe
+from pathlib import Path
 from typing import Optional
 
-import requests
+import httpx
+import sentry_sdk
 from aioprometheus.asgi.middleware import MetricsMiddleware
 from aioprometheus.asgi.starlette import metrics
 from authlib.integrations.starlette_client import OAuth, OAuthError
-from fastapi import Cookie, FastAPI, Response, WebSocket
+from fastapi import Cookie, FastAPI, File, Response, UploadFile, WebSocket
+from fastapi.responses import JSONResponse
 from fastapi.staticfiles import StaticFiles
 from pydantic import BaseModel, ValidationError
 from starlette.config import Config
 from starlette.exceptions import HTTPException as StarletteHTTPException
+from starlette.middleware.base import BaseHTTPMiddleware
 from starlette.middleware.sessions import SessionMiddleware
 from starlette.requests import Request
 from starlette.responses import HTMLResponse, RedirectResponse
+from starlette.types import ASGIApp
 
-from .broker import mqtt
-from .data.authorized import authorized
-from .websocket import ws_manager
+from turbacz.auth import (
+    JWT_EXPIRE_MINUTES,
+    create_jwt,
+    get_current_user,
+    websocket_auth,
+)
+from turbacz.broker import mqtt
+from turbacz.data.authorized import authorized
+from turbacz.data.secrets import DSN
+from turbacz.state import relay_state
+from turbacz.websocket import ws_manager
 
 logger = logging.getLogger(__name__)
 
-app = FastAPI()
+
+class CustomRequestSizeMiddleware(BaseHTTPMiddleware):
+    def __init__(self, app: ASGIApp, max_content_size: int):
+        super().__init__(app)
+        self.max_content_size = max_content_size
+
+    async def dispatch(self, request: Request, call_next: callable) -> Response:
+        content_length = int(request.headers.get("content-length", 0))
+        if content_length > self.max_content_size:
+            return Response(
+                content="Request body too large",
+                status_code=413,
+                media_type="text/plain",
+            )
+        return await call_next(request)
+
+
+MAX_REQUEST_SIZE = 10_000_000
+
+sentry_sdk.init(
+    dsn=DSN,
+    send_default_pii=True,
+    traces_sample_rate=1.0,
+)
+
+
+app = FastAPI(title="Turbacz Home Automation System", version="0.1.0")
+app.add_middleware(CustomRequestSizeMiddleware, max_content_size=MAX_REQUEST_SIZE)
 app.add_middleware(SessionMiddleware, secret_key="!secret")
 app.add_middleware(MetricsMiddleware)
 app.add_route("/metrics", metrics)
@@ -36,15 +75,13 @@ background_task_started = False
 
 app.mount("/static", StaticFiles(directory="./static", html=True), name="static")
 
+
 CONF_URL = "https://accounts.google.com/.well-known/openid-configuration"
 oauth.register(
     name="google",
     server_metadata_url=CONF_URL,
     client_kwargs={"scope": "openid email profile"},
 )
-
-with open("turbacz/data/cookies.pickle", "rb") as cookies:
-    access_cookies: dict = load(cookies)
 
 
 @app.exception_handler(StarletteHTTPException)
@@ -56,45 +93,109 @@ async def custom_http_exception_handler(request: Request, exc):
 
 @app.get("/")
 async def homepage(request: Request, access_token: Optional[str] = Cookie(None)):
-    user = request.session.get("user")
-    if access_token in access_cookies:
-        return RedirectResponse(url="/auto")
+    user = get_current_user(access_token)
+
     if user:
-        return HTMLResponse('<h1>Sio!</h1><a href="/login">login</a>')
+        return RedirectResponse(url="/auto")
+
     return HTMLResponse('<a href="/login">login</a>')
+
+
+@app.get("/login")
+async def login(request: Request):
+    redirect_uri = request.url_for("auth")
+    return await oauth.google.authorize_redirect(request, redirect_uri)
+
+
+@app.get("/auth")
+async def auth(request: Request):
+    try:
+        token = await oauth.google.authorize_access_token(request)
+
+    except OAuthError as error:
+        return HTMLResponse(f"<h1>{error.error}</h1>")
+
+    user = token.get("userinfo")
+
+    if user and user["email"] in authorized:
+        jwt_token = create_jwt(
+            {
+                "sub": user["email"],
+                "name": user.get("name"),
+            }
+        )
+
+        response = RedirectResponse(url="/auto")
+        response.set_cookie(
+            "access_token",
+            jwt_token,
+            httponly=False,
+            secure=True,  # Set to False in development if not using HTTPS
+            samesite="lax",
+            max_age=JWT_EXPIRE_MINUTES * 60,
+        )
+
+        return response
+
+    return RedirectResponse(url="/")
+
+
+@app.get("/logout")
+async def logout(response: Response):
+    response = RedirectResponse(url="/")
+    response.delete_cookie("access_token")
+
+    return response
 
 
 @app.get("/auto")
 async def main(request: Request, access_token: Optional[str] = Cookie(None)):
-    user = request.session.get("user")
-    if user and access_token in access_cookies:
+    user = get_current_user(access_token)
+
+    if user:
         with open(os.path.join("static", "index.html")) as fh:
             data = fh.read()
+
         return Response(content=data, media_type="text/html")
+
+    return RedirectResponse(url="/")
+
+
+@app.get("/heating")
+async def heating(request: Request, access_token: Optional[str] = Cookie(None)):
+    user = get_current_user(access_token)
+
+    if user:
+        with open(os.path.join("static", "heating.html")) as fh:
+            data = fh.read()
+
+        return Response(content=data, media_type="text/html")
+
     return RedirectResponse(url="/")
 
 
 @app.get("/api/temperatures")
 async def get_temperatures(request: Request, start: int, end: int, step: int):
-    response1 = requests.get(
-        "http://127.0.0.1:8428/api/v1/query_range",
-        params={
-            "start": start,
-            "end": end,
-            "query": "water_temperature",
-            "step": step,
-        },
-    )
+    async with httpx.AsyncClient() as client:
+        response1 = await client.get(
+            "http://127.0.0.1:8428/api/v1/query_range",
+            params={
+                "start": start,
+                "end": end,
+                "query": "water_temperature",
+                "step": step,
+            },
+        )
 
-    response2 = requests.get(
-        "http://127.0.0.1:8428/api/v1/query_range",
-        params={
-            "start": start,
-            "end": end,
-            "query": "pid_target",
-            "step": step,
-        },
-    )
+        response2 = await client.get(
+            "http://127.0.0.1:8428/api/v1/query_range",
+            params={
+                "start": start,
+                "end": end,
+                "query": "pid_target",
+                "step": step,
+            },
+        )
 
     if response1.status_code != 200 or response2.status_code != 200:
         return "connection not working"
@@ -117,115 +218,92 @@ async def get_temperatures(request: Request, start: int, end: int, step: int):
     return result
 
 
-@app.get("/api/heating_data")
-async def get_heating_data(request: Request):
-    response1 = requests.get(
-        "http://127.0.0.1:8428/api/v1/query?query=water_temperature"
-    )
-    response2 = requests.get("http://127.0.0.1:8428/api/v1/query?query=pid_target")
-    response3 = requests.get("http://127.0.0.1:8428/api/v1/query?query=pid_integral")
-    response4 = requests.get("http://127.0.0.1:8428/api/v1/query?query=pid_output")
-    response5 = requests.get("http://127.0.0.1:8428/api/v1/query?query=pid_multiplier")
-
-    if (
-        response1.status_code != 200
-        or response2.status_code != 200
-        or response3.status_code != 200
-        or response4.status_code != 200
-        or response5.status_code != 200
-    ):
-        return "connection not working"
-
-    heating_data = (
-        response1.json()["data"]["result"]
-        + response2.json()["data"]["result"]
-        + response3.json()["data"]["result"]
-        + response4.json()["data"]["result"]
-        + response5.json()["data"]["result"]
-    )
-
-    result = {
-        "cold": heating_data[0]["value"][1],
-        "hot": heating_data[1]["value"][1],
-        "mixed": heating_data[2]["value"][1],
-        "target": heating_data[3]["value"][1],
-        "integral": heating_data[4]["value"][1],
-        "pid": heating_data[5]["value"][1],
-        "kd": heating_data[6]["value"][1],
-        "ki": heating_data[7]["value"][1],
-        "kp": heating_data[8]["value"][1],
-    }
-
-    return result
-
-
-@app.get("/heating")
-async def heating(request: Request, access_token: Optional[str] = Cookie(None)):
-    user = request.session.get("user")
-    if user and access_token in access_cookies:
-        with open(os.path.join("static", "heating.html")) as fh:
-            data = fh.read()
-        return Response(content=data, media_type="text/html")
-    return RedirectResponse(url="/")
-
-
 @app.get("/blinds")
 async def blinds(request: Request, access_token: Optional[str] = Cookie(None)):
-    user = request.session.get("user")
-    if user and access_token in access_cookies:
+    user = get_current_user(access_token)
+
+    if user:
         with open(os.path.join("static", "blinds.html")) as fh:
             data = fh.read()
+
         return Response(content=data, media_type="text/html")
+
     return RedirectResponse(url="/")
 
 
 @app.get("/lights")
 async def lights(request: Request, access_token: Optional[str] = Cookie(None)):
-    user = request.session.get("user")
-    if user and access_token in access_cookies:
+    user = get_current_user(access_token)
+
+    if user:
         with open(os.path.join("static", "lights.html")) as fh:
             data = fh.read()
+
         return Response(content=data, media_type="text/html")
+
     return RedirectResponse(url="/")
 
 
-@app.get("/login")
-async def login(request: Request):
-    redirect_uri = request.url_for("auth")
-    return await oauth.google.authorize_redirect(request, redirect_uri)
+@app.get("/rcm")
+async def rcm_page(request: Request, access_token: Optional[str] = Cookie(None)):
+    user = get_current_user(access_token)
+
+    if not user:
+        return RedirectResponse(url="/")
+
+    with open(os.path.join("static", "relay_connection_manager.html")) as fh:
+        data = fh.read()
+
+    return Response(content=data, media_type="text/html")
 
 
-@app.get("/auth")
-async def auth(request: Request):
-    try:
-        token = await oauth.google.authorize_access_token(request)
-    except OAuthError as error:
-        return HTMLResponse(f"<h1>{error.error}</h1>")
-    user = token.get("userinfo")
-    if user:
-        request.session["user"] = dict(user)
-        if user["email"] in authorized:
-            access_token = token_urlsafe()
-            access_cookies[access_token] = user["email"]
-            with open("turbacz/data/cookies.pickle", "wb") as cookies:
-                dump(access_cookies, cookies)
-            response = RedirectResponse(url="/auto")
-            response.set_cookie("access_token", access_token, max_age=3600 * 24 * 14)
-            return response
-        else:
-            return RedirectResponse(url="/")
+@app.get("/upload")
+async def upload_page(request: Request, access_token: Optional[str] = Cookie(None)):
+    user = get_current_user(access_token)
+
+    if not user:
+        return RedirectResponse(url="/")
+
+    with open(os.path.join("static", "upload.html")) as fh:
+        data = fh.read()
+
+    return Response(content=data, media_type="text/html")
 
 
-@app.get("/logout")
-async def logout(
-    request: Request, response: Response, access_token: Optional[str] = Cookie(None)
+@app.post("/upload/{device}")
+async def upload_firmware(
+    request: Request,
+    device: str,
+    file: UploadFile = File(...),
+    access_token: Optional[str] = Cookie(None),
 ):
-    access_cookies.pop(access_token)
-    with open("turbacz/data/cookies.pickle", "wb") as cookies:
-        dump(access_cookies, cookies)
-    request.session.pop("user", None)
-    response.delete_cookie(key="access_token")
-    return RedirectResponse(url="/")
+    user = get_current_user(access_token)
+
+    if not user:
+        return RedirectResponse(url="/")
+
+    print("Uploading firmware for device:", device)
+
+    if device not in ["switch", "relay", "root"]:
+        return JSONResponse(
+            {"status": "error", "reason": "unknown device"}, status_code=400
+        )
+
+    try:
+        save_path = Path(f"static/data/{device}/firmware.bin")
+        save_path.parent.mkdir(parents=True, exist_ok=True)
+
+        contents = await file.read()
+
+        with open(save_path, "wb") as buffer:
+            buffer.write(contents)
+
+        logger.info(f"Successfully saved {len(contents)} bytes to {save_path}")
+        return JSONResponse({"status": "ok", "device": device, "size": len(contents)})
+
+    except Exception as e:
+        logger.error(f"Error uploading firmware: {e}", exc_info=True)
+        return JSONResponse({"status": "error", "reason": str(e)}, status_code=500)
 
 
 class BlindRequest(BaseModel):
@@ -244,14 +322,22 @@ async def set_blind(req: BlindRequest):
 
 
 @app.websocket("/blinds/ws/{client_id}")
-async def websocket_blinds(websocket: WebSocket, access_token=Cookie()):
+async def websocket_blinds(websocket: WebSocket):
+    user = await websocket_auth(websocket)
+
+    if not user:
+        await websocket.close(code=1008)
+
+        return
+
     await ws_manager.connect(websocket)
+
     mqtt.publish("/blind/cmd", "S")
 
     async def receive_command(websocket: WebSocket):
         async for cmd in websocket.iter_json():
             try:
-                req = BlindRequest.parse_obj(cmd)
+                req = BlindRequest.model_validate(cmd)
             except ValidationError as err:
                 logger.error("Cannot parse %s %s", cmd, err)
                 continue
@@ -260,12 +346,18 @@ async def websocket_blinds(websocket: WebSocket, access_token=Cookie()):
                 "/blind/cmd", f"{chr(int(req.blind[1])+96)}{req.position}"
             )
 
-    if access_token in access_cookies:
-        await receive_command(websocket)
+    await receive_command(websocket)
 
 
 @app.websocket("/heating/ws/{client_id}")
-async def websocket_heating(websocket: WebSocket, access_token=Cookie()):
+async def websocket_heating(websocket: WebSocket):
+    user = await websocket_auth(websocket)
+
+    if not user:
+        await websocket.close(code=1008)
+
+        return
+
     await ws_manager.connect(websocket)
 
     async def receive_command(websocket: WebSocket):
@@ -273,35 +365,96 @@ async def websocket_heating(websocket: WebSocket, access_token=Cookie()):
             logger.debug("putting %s in command queue", cmd)
             mqtt.client.publish("/heating/cmd", cmd)
 
-    if access_token in access_cookies:
-        await receive_command(websocket)
+    await receive_command(websocket)
 
 
 @app.websocket("/lights/ws/{client_id}")
-async def websocket_lights(websocket: WebSocket, access_token=Cookie()):
+async def websocket_lights(websocket: WebSocket):
+    user = await websocket_auth(websocket)
+
+    if not user:
+        await websocket.close(code=1008)
+
+        return
+
     await ws_manager.connect(websocket)
-    mqtt.publish("/switch/1/cmd", "S")
+
+    current_states = relay_state.get_all()
+    for light_id, state in current_states.items():
+        await ws_manager.send_personal_message(
+            {"id": light_id, "state": state}, websocket
+        )
 
     async def receive_command(websocket: WebSocket):
         async for cmd in websocket.iter_json():
             try:
-                chg = SwitchChange.parse_obj(cmd)
+                chg = SwitchChange.model_validate(cmd)
             except ValidationError as err:
                 logger.error("Cannot parse %s %s", cmd, err)
                 continue
             logger.debug("putting %s in command queue", cmd)
-            print(f"{chg.id}{chg.state}")
-            mqtt.client.publish("/switch/1/cmd", f"{chg.id}{chg.state}")
+            chg.id = int(chg.id[1:])
 
-    if access_token in access_cookies:
-        await receive_command(websocket)
+            match chg.id // 8:
+                case 0:
+                    topic = "/switch/1/cmd"
+
+                case 1:
+                    topic = "/relay/cmd/1074130365"
+
+                case 2:
+                    topic = "/relay/cmd/1074122133"
+
+                case _:
+                    raise ValueError("Invalid light id")
+
+            chg.id = chr(chg.id % 8 + 97)
+
+            print(topic, f"{chg.id}{chg.state}")  # Debug print
+            mqtt.client.publish(topic, f"{chg.id}{chg.state}")
+
+    await receive_command(websocket)
+
+
+@app.websocket("/rcm/ws/{client_id}")
+async def websocket_rcm(websocket: WebSocket):
+    user = await websocket_auth(websocket)
+
+    if not user:
+        await websocket.close(code=1008)
+
+        return
+
+    await ws_manager.connect(websocket)
+
+    with open("turbacz/data/connections.json", "r", encoding="utf-8") as f:
+        rcm_config = json.load(f)
+        print(rcm_config)
+
+        await ws_manager.send_personal_message(rcm_config, websocket)
+
+    async def receive_command(websocket: WebSocket):
+        async for cmd in websocket.iter_json():
+            logger.debug("putting %s in command queue", cmd)
+
+            with open("turbacz/data/connections.json", "w", encoding="utf-8") as f:
+                json.dump(cmd, f, ensure_ascii=False, indent=2)
+
+            for connection in list(cmd["connections"].keys()):
+                cmd["connections"][connection[:10]] = cmd["connections"].pop(connection)
+
+            print(f"Updated connections: {cmd['connections']}")
+
+            mqtt.client.publish("/switch/cmd/root", cmd["connections"])
+
+    await receive_command(websocket)
 
 
 def start():
     import uvicorn
 
     logging.basicConfig(level=logging.DEBUG)
-    uvicorn.run(app, host="127.0.0.1", port=8000)
+    uvicorn.run(app, host="127.0.0.1", port=8002)
 
 
 if __name__ == "__main__":
