@@ -1,13 +1,72 @@
 var wsId = Math.floor(Math.random() * 2000000000)
-var ws = new WebSocket("ws://127.0.0.1:8000/blinds/ws/" + wsId);
+var ws = null
+var reconnectTimeout = null
+var reconnectDelay = 1000
+var maxReconnectDelay = 30000
+var isReconnecting = false
 
-ws.onmessage = function (event) {
-	var msg = JSON.parse(event.data)
-	console.info(msg)
-	$("#" + msg.blind).slider("value", 999 - msg.current_position)
+
+function getCookie(cname) {
+	let name = cname + "=";
+	let decodedCookie = decodeURIComponent(document.cookie);
+	let ca = decodedCookie.split(';');
+	for (let i = 0; i < ca.length; i++) {
+		let c = ca[i];
+		while (c.charAt(0) == ' ') {
+			c = c.substring(1);
+		}
+		if (c.indexOf(name) == 0) {
+			return c.substring(name.length, c.length);
+		}
+	}
+	return "";
+}
+
+function connectWebSocket() {
+	if (isReconnecting) return
+	isReconnecting = true
+
+	console.log('Connecting WebSocket...')
+	auth_token = getCookie("access_token")
+	ws = new WebSocket(`wss://${window.location.host}/blinds/ws/` + wsId + `?token=` + auth_token)
+
+	ws.onopen = function () {
+		console.log('WebSocket connected!')
+		isReconnecting = false
+		reconnectDelay = 1000
+	}
+
+	ws.onmessage = function (event) {
+		var msg = JSON.parse(event.data)
+		console.info(msg)
+		$("#" + msg.blind).slider("value", 999 - msg.current_position)
+	}
+
+	ws.onerror = function (error) {
+		console.error('WebSocket error:', error)
+	}
+
+	ws.onclose = function (event) {
+		console.log('WebSocket disconnected')
+		isReconnecting = false
+
+		if (reconnectTimeout) {
+			clearTimeout(reconnectTimeout)
+		}
+
+		console.log(`Reconnecting in ${reconnectDelay / 1000}s...`)
+		reconnectTimeout = setTimeout(function () {
+			connectWebSocket()
+			reconnectDelay = Math.min(reconnectDelay * 2, maxReconnectDelay)
+		}, reconnectDelay)
+	}
 }
 
 $(function () {
+	// Initialize connection
+	connectWebSocket()
+
+	// Initialize sliders
 	$(".flex-container > span").each(function () {
 		var value = parseInt($(this).text(), 10)
 		$(this).empty().slider({
@@ -18,20 +77,51 @@ $(function () {
 			orientation: "vertical",
 			stop: function (event, ui) {
 				console.info(event.target.id, ui.value)
-				ws.send(JSON.stringify({ "blind": event.target.id, "position": parseInt(999 - ui.value) }))
+
+				// Check if WebSocket is connected
+				if (!ws || ws.readyState !== WebSocket.OPEN) {
+					console.log('WebSocket not connected, trying to reconnect...')
+					connectWebSocket()
+					return
+				}
+
+				try {
+					ws.send(JSON.stringify({ "blind": event.target.id, "position": parseInt(999 - ui.value) }))
+				} catch (e) {
+					console.error('Failed to send:', e)
+					connectWebSocket()
+				}
 			}
 		})
 	})
 })
 
+// Check connection every 30 seconds
+setInterval(function () {
+	if (!ws || ws.readyState === WebSocket.CLOSED) {
+		console.log('WebSocket closed, reconnecting...')
+		connectWebSocket()
+	}
+}, 30000)
+
+// Handling visibility API
+document.addEventListener('visibilitychange', function () {
+	if (!document.hidden) {
+		if (!ws || ws.readyState !== WebSocket.OPEN) {
+			console.log('Tab visible again, checking connection...')
+			connectWebSocket()
+		}
+	}
+})
+
 function openNav() {
-	document.getElementById("sidenav").style.width = "160px";
-	document.getElementById("main").style.marginLeft = "160px";
-	document.getElementById("openbtn").style.visibility = "hidden";
+	document.getElementById("sidenav").style.width = "160px"
+	document.getElementById("main").style.marginLeft = "160px"
+	document.getElementById("openbtn").style.visibility = "hidden"
 }
 
 function closeNav() {
-	document.getElementById("sidenav").style.width = "0";
-	document.getElementById("main").style.marginLeft = "0";
-	document.getElementById("openbtn").style.visibility = "visible";
-} 
+	document.getElementById("sidenav").style.width = "0"
+	document.getElementById("main").style.marginLeft = "0"
+	document.getElementById("openbtn").style.visibility = "visible"
+}
